+++ R/scatter_plotly.R
@@ -4,14 +4,13 @@
 #' Produces a scatter plot from a 'locus' class object using plotly.
 #'
 #' @param loc Object of class 'locus' to use for plot. See [locus].
-#' @param index_snp Specifies index SNP to be shown in a different colour and
-#'   symbol. Defaults to the SNP with the lowest p-value. Set to `NULL` to not
-#'   show this.
+#' @param index_snp Specifies index SNP or a vector of SNPs to be shown in a
+#'   different colour and symbol. Defaults to the SNP with the lowest p-value.
+#'   Set to `NULL` to not show this.
 #' @param pcutoff Cut-off for p value significance. Defaults to p = 5e-08. Set
 #'   to `NULL` to disable.
-#' @param chromCol Colour for normal points if `LD` is `FALSE` when the locus
-#'   object is made.
-#' @param sigCol Colour for significant points if `LD` is `FALSE`.
+#' @param scheme Vector of 3 colours if LD is not shown: 1st = normal points,
+#'   2nd = colour for significant points, 3rd = index SNP(s).
 #' @param xlab x axis title.
 #' @param ylab y axis title.
 #' @param yzero Logical whether to force y axis limit to include y=0.
@@ -25,16 +24,23 @@
 #' @param recomb_col Colour for recombination rate line if recombination rate
 #'   data is present. Set to `NA` to hide the line. See [link_recomb()] to add
 #'   recombination rate data.
+#' @param eqtl_gene Column name in `loc$data` for eQTL genes.
+#' @param beta Optional column name for beta coefficient to display upward
+#'   triangles for positive beta and downward triangles for negative beta
+#'   (significant SNPs only).
+#' @param add_hover Optional vector of column names in `loc$data` to add to the
+#'   plotly hover text for scatter points.
+#' @param showlegend Logical whether to show a legend for the scatter points.
 #' @return A `plotly` scatter plot.
 #' @seealso [locus()] [locus_plotly()]
-#' @importFrom plotly add_trace
+#' @importFrom plotly add_trace plotly_build
+#' @importFrom stats relevel
 #' @export
 #' 
 scatter_plotly <- function(loc,
                            index_snp = loc$index_snp,
                            pcutoff = 5e-08,
-                           chromCol = 'royalblue',
-                           sigCol = 'red',
+                           scheme = c('grey', 'dodgerblue', 'red'),
                            xlab = NULL,
                            ylab = NULL,
                            yzero = (loc$yvar == "logP"),
@@ -43,9 +49,15 @@ scatter_plotly <- function(loc,
                                          'orange', 'red', 'purple'),
                            marker_outline = "black",
                            marker_size = 7,
-                           recomb_col = "blue") {
+                           recomb_col = "blue",
+                           eqtl_gene = NULL,
+                           beta = NULL,
+                           add_hover = NULL,
+                           showlegend = TRUE) {
   if (!inherits(loc, "locus")) stop("Object of class 'locus' required")
-  if (is.null(loc$data)) stop("No data points, only gene tracks")
+  if (is.null(loc$data)) stop("No SNPs/data points", call. = FALSE)
+  
+  .call <- match.call()
   data <- loc$data
   if (is.null(xlab)) xlab <- paste("Chromosome", loc$seqname, "(Mb)")
   if (is.null(ylab)) {
@@ -57,36 +69,60 @@ scatter_plotly <- function(loc,
     if (showLD & hasLD) {
       data$bg <- cut(data$ld, -1:6/5, labels = FALSE)
       data$bg[is.na(data$bg)] <- 1L
-      data$bg[data[, loc$labs] == index_snp] <- 7L
+      data$bg[data[, loc$labs] %in% index_snp] <- 7L
       data <- data[order(data$bg), ]
-      LD_scheme <- rep_len(LD_scheme, 7 - is.null(index_snp))
+      scheme <- rep_len(LD_scheme, 7 - is.null(index_snp))
       data$bg <- factor(data$bg, levels = 1:7,
                         labels = c("unknown", "0.0 - 0.2", "0.2 - 0.4",
                                    "0.4 - 0.6", "0.6 - 0.8", "0.8 - 1.0",
                                    "index"))
       leg <- list(title = list(text = "Linkage r<sup>2</sup>"))
+    } else if (!is.null(eqtl_gene)) {
+      bg <- data[, eqtl_gene]
+      bg[data[, loc$p] > pcutoff] <- "ns"
+      data$bg <- relevel(factor(bg, levels = unique(bg)), "ns")
+      if (is.null(.call$scheme)) scheme <- eqtl_scheme(nlevels(data$bg))
     } else {
+      # default colours
       showLD <- FALSE
-      data$bg <- chromCol
-      if (loc$yvar == "logP") data$bg[data[, loc$p] < pcutoff] <- sigCol
-      data$bg[data[, loc$labs] == index_snp] <- "purple"
-      LD_scheme <- c(chromCol, sigCol, "purple")
-      data$bg <- factor(data$bg, levels = LD_scheme,
-                        labels = c("ns", paste("P <", pcutoff), "index"))
+      data$bg <- scheme[1]
+      if (loc$yvar == "logP") data$bg[data[, loc$p] < pcutoff] <- scheme[2]
+      data$bg[data[, loc$labs] %in% index_snp] <- scheme[3]
+      data$bg <- factor(data$bg, levels = scheme,
+                        labels = c("ns", paste("P <", signif(pcutoff, 3)), "index"))
+    }
+  }
+  if (!is.null(beta)) {
+    # beta symbols
+    data[, beta] <- signif(data[, beta], 3)
+    symbol <- as.character(sign(data[, beta]))
+    ind <- data[, loc$p] > pcutoff
+    symbol[ind] <- "ns"
+    data$symbol <- factor(symbol, levels = c("ns", "1", "-1"),
+                          labels = c(" ", "up", "down"))
+    # labels = c(" ", "&#946; > 0", "&#946; < 0")
+    symbols <- c(21L, 24L, 25L)
+    data$size <- 1L
+    data$size[!ind] <- 2L
+    sizes <- if (sum(!ind) == 0) c(50, 50) else c(50, 100)
+    leg <- list(traceorder = "reversed")
+  } else {
+    if (is.null(eqtl_gene)) {
+      # default plot
+      data$symbol <- data$bg
+      symbols <- c(rep("circle", length(scheme) -1), "diamond")
+    } else {
+      # eqtl gene only
+      data$symbol <- 1L
+      symbols <- "circle"  
     }
   }
   
   # scatter plotly
   recomb <- !is.null(loc$recomb) & !is.na(recomb_col)
   
-  pch <- rep(21L, nrow(data))
-  pch[data[, loc$labs] == index_snp] <- 23L
-  if ("pch" %in% colnames(data)) pch <- data$pch
-  col <- "black"
-  if ("col" %in% colnames(data)) col <- data$col
-  
   xlim <- loc$xrange / 1e6
-  xext <- diff(xlim) * 0.05
+  xext <- diff(xlim) * 0.01
   xlim <- xlim + c(-xext, xext)
   
   ylim <- range(data[, loc$yvar], na.rm = TRUE)
@@ -98,57 +134,133 @@ scatter_plotly <- function(loc,
   hovertext <- paste0(data[, loc$labs], "<br>Chr ",
                       data[, loc$chrom], ": ", data[, loc$pos],
                       "<br>P = ", signif(data[, loc$p], 3))
-  ylim2 <- c(-2, 102)
+  add_hover <- c(beta, eqtl_gene, add_hover)
+  if (!is.null(add_hover)) {
+    for (i in add_hover) {
+      hovertext <- paste0(hovertext, "<br>", i, ": ", data[, i])
+    }
+  }
+  
+  hline <- list(type = "line",
+                line = list(width = 1, color = '#999999', dash = 'dash'),
+                x0 = 0, x1 = 1, y0 = -log10(pcutoff), y1 = -log10(pcutoff),
+                xref = "paper", layer = "below")
   
   if (!recomb) {
-    # standard plotly
-    plot_ly(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
-            color = data$bg, colors = LD_scheme,
-            marker = list(size = marker_size, opacity = 0.8,
-                          line = list(width = 1, color = marker_outline)),
-            text = hovertext,
-            hoverinfo = 'text',
-            type = "scattergl", mode = "markers") %>%
+    if (is.null(beta)) {
+      # standard plotly
+      p <- plot_ly(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
+                   color = data$bg, colors = scheme,
+                   symbol = data$symbol, symbols = symbols,
+                   marker = list(size = marker_size, opacity = 0.8,
+                                 line = list(width = 1, color = marker_outline)),
+                   text = hovertext, hoverinfo = 'text',
+                   key = data[, loc$labs],
+                   showlegend = showlegend,
+                   source = "plotly_locus",
+                   type = "scattergl", mode = "markers")
+    } else {
+      # beta shapes
+      p <- plot_ly(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
+                   color = data$bg, colors = scheme,
+                   symbol = data$symbol, symbols = symbols,
+                   size = data$size, sizes = sizes,
+                   marker = list(opacity = 0.8,
+                                 line = list(width = 1, color = marker_outline)),
+                   text = hovertext, hoverinfo = 'text',
+                   key = data[, loc$labs],
+                   showlegend = showlegend,
+                   source = "plotly_locus",
+                   type = "scattergl", mode = "markers")
+    }
+    p <- p %>%
       plotly::layout(xaxis = list(title = xlab,
                                   ticks = "outside",
+                                  zeroline = FALSE, showgrid = FALSE,
                                   range = as.list(xlim)),
                      yaxis = list(title = ylab,
                                   ticks = "outside",
-                                  showline = TRUE, range = ylim),
-                     legend = leg,
-                     showlegend = showLD | !is.null(pcutoff)) %>%
-      plotly::config(displaylogo = FALSE)
-    # modeBarButtonsToRemove = c("zoom2d", "pan2d", "select2d",
-    #                            "lasso2d", "zoomIn2d", "zoomOut2d", 
-    #                            "autoScale2d", "toggleHover"))
+                                  fixedrange = TRUE,
+                                  showline = TRUE,
+                                  range = ylim),
+                     shapes = hline, legend = leg, dragmode = "pan")
   } else {
     # double y axis with recombination
-    plot_ly() %>%
-      add_trace(x = loc$recomb$start / 1e6, y = loc$recomb$value,
-                hoverinfo = "none", colors = LD_scheme,  # colors must go here
-                name = "recombination", yaxis = "y2",
-                line = list(color = recomb_col),
-                mode = "lines", type = "scattergl", showlegend = FALSE) %>%
-      add_trace(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
-                color = data$bg,
-                marker = list(size = marker_size, opacity = 0.8,
-                              line = list(width = 1, color = marker_outline)),
-                text = hovertext,
-                hoverinfo = 'text',
-                type = "scattergl", mode = "markers") %>%
+    ylim2 <- c(-2, 102)
+    if (is.null(beta)) {
+      # standard plotly
+      p <- plot_ly(source = "plotly_locus") %>%
+        # recombination line
+        add_trace(x = loc$recomb$start / 1e6, y = loc$recomb$value,
+                  hoverinfo = "none", colors = scheme,  # colors must go here
+                  symbols = symbols,
+                  name = "recombination", yaxis = "y2",
+                  line = list(color = recomb_col, width = 1.5),
+                  mode = "lines", type = "scattergl", showlegend = FALSE) %>%
+        # scatter plot
+        add_trace(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
+                  color = data$bg,
+                  symbol = data$symbol,
+                  marker = list(size = marker_size, opacity = 0.8,
+                                line = list(width = 1, color = marker_outline)),
+                  text = hovertext, hoverinfo = 'text', key = data[, loc$labs],
+                  showlegend = showlegend,
+                  type = "scattergl", mode = "markers")
+    } else {
+      # beta shapes
+      p <- plot_ly(source = "plotly_locus") %>%
+        # recombination line
+        add_trace(x = loc$recomb$start / 1e6, y = loc$recomb$value,
+                  hoverinfo = "none", colors = scheme,  # colors must go here
+                  symbols = symbols, sizes = sizes,
+                  name = "recombination", yaxis = "y2",
+                  line = list(color = recomb_col, width = 1.5),
+                  mode = "lines", type = "scattergl", showlegend = FALSE) %>%
+        # scatter plot
+        add_trace(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
+                  color = data$bg,
+                  symbol = data$symbol,
+                  size = data$size,
+                  marker = list(opacity = 0.8,
+                                line = list(width = 1, color = marker_outline)),
+                  text = hovertext, hoverinfo = 'text', key = data[, loc$labs],
+                  showlegend = showlegend,
+                  type = "scattergl", mode = "markers")
+    }
+    p <- p %>%
       plotly::layout(xaxis = list(title = xlab,
                                   ticks = "outside",
+                                  zeroline = FALSE,
                                   range = as.list(xlim)),
                      yaxis = list(title = ylab,
-                                  ticks = "outside",
-                                  showline = TRUE, range = ylim),
+                                  ticks = "outside", showgrid = FALSE,
+                                  showline = TRUE, fixedrange = TRUE,
+                                  range = ylim),
                      yaxis2 = list(overlaying = "y", side = "right",
                                    title = "Recombination rate (%)",
                                    ticks = "outside", showgrid = FALSE,
-                                   showline = TRUE,
+                                   showline = TRUE, fixedrange = TRUE,
                                    zeroline = FALSE, range = ylim2),
-                     legend = c(leg, x = 1.1, y = 1),
-                     showlegend = showLD | !is.null(pcutoff)) %>%
-      plotly::config(displaylogo = FALSE)
+                     shapes = hline,
+                     legend = c(leg, x = 1.1, y = 1), showlegend = TRUE)
   }
+  p <- p %>%
+    plotly::config(displaylogo = FALSE,
+                   modeBarButtonsToRemove = c("select2d", "lasso2d",
+                                              "autoScale2d", "resetScale2d",
+                                              "hoverClosest", "hoverCompare"))
+  
+  if (hasLD) suppressWarnings(plotly_build(p)) else p
+}
+
+
+#' @importFrom grDevices rainbow
+eqtl_scheme <- function(n) {
+  if (n < 8) {
+    scheme <- c('grey', 'purple', 'green3', 'orange', 'royalblue', 'red',
+                'cyan')[1:n]
+  } else {
+    scheme <- c('grey', rainbow(n -1))
+  }
+  scheme
 }